Moves the url replacement methods to x_module.ModuleSystem
so that they are available from the xblock runtime via LmsModuleSystem or PreviewModuleSystem.
